Fixed notification when Gemfile.lock not exists

+++ build.sh
@@ -1,5 +1,9 @@
  alias set_gemfile='export BUNDLE_GEMFILE="`pwd`/Gemfile"'
- rm Gemfile.lock
+ if [ -e "Gemfile.lock" ] 
+ then
+ 	echo "Removing Gemfile.lock..."
+ 	rm Gemfile.lock
+ fi
  bundle check || bundle install
  bundle exec rake test_app
  cd api; set_gemfile; rm Gemfile.lock; bundle install; bundle exec rspec spec
